IP based unique visit counter
This is a unique PHP page visit counter. The hit counter value is
maintained in a file count.txt. Prior to modifying the counter value IP
of the user is checked for verifying the identity. If IP is not already
present in the ip.txt file, the IP is added to the file and the hit
counter is incremented. Updated counter value is written back to file.